// src/app/api/model/plannedPlantationRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PlannedPlantationRequestDto { 
    /**
     * ID del terreno asociado (opcional)
     */
    landId?: number;
    /**
     * Fecha prevista de la plantación
     */
    plannedDate: string;
    /**
     * Fecha efectiva de la plantación (opcional)
     */
    effectiveDate?: string;
    /**
     * Cantidad mínima de árboles a plantar
     */
    minTrees: number;
    /**
     * Cantidad óptima de árboles a plantar (opcional)
     */
    optimalTrees?: number;
    /**
     * Cantidad máxima de árboles a plantar (opcional)
     */
    maxTrees?: number;
    isActive?: boolean;
}